requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
twilio
google-cloud-secret-manager==2.19.0